Improving new Ionic scaffolding; edition of events; some fixes

+++ src/pages/event-form/event-form.ts
@@ -1,19 +1,24 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams, ToastController,ViewController } from 'ionic-angular';
+import { NavController, NavParams, ToastController, ViewController } from 'ionic-angular';
 
 import { Event } from '../../entities/event';
 
 import { AuthService } from '../../services/auth.service';
 import { EventService } from '../../services/event.service';
 
-@IonicPage()
 @Component({
   selector: 'page-event-form',
   templateUrl: 'event-form.html',
 })
 export class EventFormPage {
 
+  firebaseObservable: any;
   event: Event = new Event();
+  eventId: string;
+  originalEvent: Event = new Event();
+  editMode: Boolean = false;
+  title: string = "Create event";
+  eventModified: boolean = false;
   userId: string;
   disciplines = [
     {value: 'swimming', label: 'Swimming'},
@@ -28,35 +33,84 @@ export class EventFormPage {
     private toastCtrl: ToastController,
     private auth: AuthService, 
     private eventService: EventService ) {
+      
+      this.editMode = Boolean(navParams.get('event'));
+
+      if(this.editMode) {
+        this.event = navParams.get('event');
+        this.eventId = navParams.get('eventId');
+        this.originalEvent = JSON.parse(JSON.stringify(this.event));
+        this.title = "Edit event";
+      }
   }
 
   ionViewDidLoad() {
-    this.auth.user.subscribe( user => {
-      this.userId = user.uid;
+    this.firebaseObservable = this.auth.user.subscribe( (user) => {
+      if(user) {
+        this.userId = user.uid;
+      }
     });
   }
 
-  createEvent() {
+  ionViewWillUnload() {
+    this.firebaseObservable.unsubscribe();
+  }
 
-    this.eventService.createEvent(this.event).then( (event) => {
+  ionViewWillLeave() {
+    if(!this.eventModified && this.editMode) {
+      Object.assign(this.event, this.originalEvent);
+    }
+  }
+
+  saveEvent() {
+
+    if(!this.editMode) {
+      this.event.owner = this.userId;
+      this.eventService.createEvent(this.event).then( (event) => {
+        
+        let toast = this.toastCtrl.create({
+          message: 'Event created',
+          duration: 3000,
+          position: 'bottom',
+          showCloseButton: true,
+          closeButtonText: 'UNDO'
+        });
       
-      let toast = this.toastCtrl.create({
-        message: 'Event created',
-        duration: 3000,
-        position: 'bottom',
-        showCloseButton: true,
-        closeButtonText: 'UNDO'
-      });
-    
-      toast.onDidDismiss((data, role) => {
-        if (role == "close") {
-          this.eventService.deleteEvent(event.id);
-        }
-      });
-    
-      toast.present();
+        toast.onDidDismiss((data, role) => {
+          if (role == "close") {
+            this.eventService.deleteEvent(event.id);
+          }
+        });
+      
+        toast.present();
 
-    });
+      });
+    } else {
+
+      this.eventModified = true;
+
+      this.eventService.updateEvent(this.eventId, this.event).then( () => {
+      
+        let toast = this.toastCtrl.create({
+          message: 'Event modified',
+          duration: 3000,
+          position: 'bottom',
+          showCloseButton: true,
+          closeButtonText: 'UNDO'
+        });
+      
+        toast.onDidDismiss((data, role) => {
+          if (role == "close") {
+            this.eventService.updateEvent(this.eventId, this.originalEvent).then( () => {
+              Object.assign(this.event, this.originalEvent);
+            });
+          }
+        });
+      
+        toast.present();
+  
+      });
+    }
 
     this.viewCtrl.dismiss();
   }
